+++ prettier.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  parser: "flow",
+  bracketSpacing: true,
+  jsxBracketSameLine: true,
+  semi: false,
+  trailingComma: "es5",
+}